Add datapackage stuff to seperate file
