+++ main.js
@@ -7,6 +7,8 @@ var Fs = require('fs');
 var XLSX = require('xlsx');
 var ipc = require('ipc');
 
+var datapackage = require('./datapackage')
+
 // Report crashes to our server.
 require('crash-reporter').start();
 
@@ -92,7 +94,7 @@ app.on('ready', function() {
         },
         {
           label: 'Export as Datapackage',
-          click: function() { exportDatapackage(); }
+          click: function() { datapackage.exportDatapackage(); }
         },
         {
           label: 'Validate',
@@ -279,58 +281,4 @@ function importExcel() {
 function validateFile() {
   window = BrowserWindow.getFocusedWindow();
   window.webContents.send('validate');
-}
-
-function datapackageJson(data) {
-  data.keywords = data.keywords.split(",")
-  data.resources = [
-    {
-      "name": data.name,
-      "path": "data/" + data.name + ".csv",
-      "mediatype": "text/csv"
-    }
-  ]
-  return data
-}
-
-function generateDatapackage(fileName, data, csv) {
-  zip = new require('node-zip')();
-  zip.file('datapackage.json', JSON.stringify(data));
-  zip.file('data/' + data.name + '.csv', csv);
-  zipData = zip.generate({base64:false,compression:'DEFLATE'});
-  Fs.writeFileSync(fileName, zipData, 'binary');
-}
-
-function exportDatapackage() {
-  var window = BrowserWindow.getFocusedWindow();
-
-  datapackage = new BrowserWindow({width: 450, height: 600, 'always-on-top': true});
-  datapackage.loadUrl('file://' + __dirname + '/datapackage.html');
-
-  datapackage.on('closed', function() {
-    datapackage = null;
-  });
-
-  ipc.once('sendDatapackage', function(e, data) {
-    var data = datapackageJson(data)
-
-    Dialog.showSaveDialog({
-      filters: [
-        { name: 'text', extensions: ['zip'] }
-      ],
-      defaultPath: 'datapackage.zip'
-    }, function (fileName) {
-      if (fileName === undefined) return;
-      datapackage.close();
-      window.webContents.send('getCSV');
-
-      ipc.once('sendCSV', function(e, csv) {
-        generateDatapackage(fileName, data, csv)
-      });
-    });
-  });
-
-  ipc.once('datapackageCanceled', function() {
-    datapackage.close();
-  });
 }
